refactor: ディレクトリ構成を修正 参考: https://qiita.com/sueken/items/87093e5941bfbc09bea8
