+++ cmd/gateway/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"gateway/internal/apps/gateway"
+)
+
+func main() {
+	gateway.Gateway()
+}
