+++ internal/apps/gateway/gateway_internal_test.go
@@ -1,0 +1,9 @@
+package gateway
+
+import (
+	"testing"
+)
+
+func TestGateWay(t *testing.T) {
+	t.Skip("skip main...")
+}
