+++ pkg/lookuptable/lookuptable.go
