+++ pkg/lookuptable/lookuptable_internal_test.go
